Frontend Implementation and backend Fixes
Beginning work on frontend. Currently SPA that allows searching for full artist name and returning results to unformatted component.

Ran into bug on backend when searching for "E-40", it seems there's a problem with my logic when special characters exist. Making adjustment to func saveArtist() to check if artist exists.